refactor: naming
RISDEV-7770
