+++ frontend/src/views/verkuendungen/verkuendungDetail/expressionenErzeugen/RisZielnormExpressionsTable.vue
